Added support for adding bills with syntax `add bill [name] [amount] [dueDate]`

+++ Scanner/Tokens/NumberToken.cs
@@ -4,8 +4,8 @@ namespace BudgetCLI.Scanner.Tokens
 {
     public class NumberToken : BudgetTokenBase
     {
-        public float Value { get; set; }
-        public NumberToken(string rawToken, float value) : base(rawToken, BudgetTokenEnum.NUMBER)
+        public decimal Value { get; set; }
+        public NumberToken(string rawToken, decimal value) : base(rawToken, BudgetTokenEnum.NUMBER)
         {
             Value = value;
         }
